+++ src/main/java/org/camunda/bpm/servlets/BeginContractServlet.java
@@ -17,8 +17,9 @@ import org.camunda.bpm.engine.impl.util.json.JSONObject;
 import org.camunda.bpm.engine.runtime.ProcessInstance;
 
 public class BeginContractServlet extends HttpServlet {
+	
 
-	public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
+	public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
 
 		ProcessEngine processEngine = ProcessEngines.getDefaultProcessEngine();
 		RuntimeService runtimeService = processEngine.getRuntimeService();
@@ -28,43 +29,16 @@ public class BeginContractServlet extends HttpServlet {
 
 		JSONObject jsonObj = new JSONObject(request.getParameter("jsonObj"));
 
-		String CustomerType = (String) jsonObj.get("customertype");
-		String phone = (String) jsonObj.get("phone");
-		String bvisProcessId = (String) jsonObj.get("process_id");
-		String endDate = (String) jsonObj.get("enddate");
-		String startDate = (String) jsonObj.get("startdate");
-		String carType = (String) jsonObj.get("vehicleclass");
-		String contract = (String) jsonObj.get("insurance");
+		String name = (String) jsonObj.get("name");
+		
 
-		map.put("CustomerType", CustomerType);
-		map.put("phone", phone);
-		map.put("bvisProcessId", bvisProcessId);
-		map.put("endDate", endDate);
-		map.put("startDate", startDate);
-		map.put("carType", carType);
-		map.put("contract", contract);
-
-		if (CustomerType.equals("private")) {
-
-			String firstName = (String) jsonObj.get("firstname");
-			String lastName = (String) jsonObj.get("lastname");
-			String birthdate = (String) jsonObj.get("birthdate");
-			String driverLicenseId = (String) jsonObj.get("licenceid");
-			String driverLicenseDate = (String) jsonObj.get("licenceexp");
-
-			map.put("firstName", firstName);
-			map.put("lastName", lastName);
-			map.put("birthdate", birthdate);
-			map.put("driverLicenseId", driverLicenseId);
-			map.put("driverLicenseDate", driverLicenseDate);
-
-		} else {
-			String name = (String) jsonObj.get("companyname");
-			map.put("name", name);
-		}
+		map.put("name", name);
+		
+		System.out.println(map.get(name));
+		
 		out.close();
 		ProcessInstance processInstance;
-		System.out.println("bvisProcessId in CreateContractServlet: " + bvisProcessId);
+		
 
 		processInstance = runtimeService.startProcessInstanceByMessage("instantiationMessageContract", map);
 	}
