+++ src/main/webapp/WEB-INF/web.xml
@@ -1,15 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <web-app xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://java.sun.com/xml/ns/javaee" xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd" version="3.0">
-  <servlet>
-    <description></description>
-    <display-name>CreateContractProcessServlet</display-name>
-    <servlet-name>CreateContractProcessServlet</servlet-name>
-    <servlet-class>org.camunda.bpm.servlet.CreateContractProcessServlet</servlet-class>
-  </servlet>
-  <servlet-mapping>
-    <servlet-name>CreateContractProcessServlet</servlet-name>
-    <url-pattern>/CreateContractProcessServlet</url-pattern>
-  </servlet-mapping>
+  
   <servlet>
     <description></description>
     <display-name>BeginContractServlet</display-name>
